pyproject.toml
[tool.poetry]
name = "web3indexer"
version = "0.1.0"
description = "An ethereum virtual machine blockchain indexer"
authors = ["Jumbo <Jummbo@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
web3 = "^5.28.0"
structlog = "^21.5.0"
pymongo = "^4.0.2"
pydantic = "^1.9.0"
kafka-python = "^2.0.2"

[tool.poetry.scripts]
index = 'web3indexer.indexer:main'
producer = 'web3indexer.producer:main'
consumer = 'web3indexer.consumer:main'

[tool.poetry.dev-dependencies]
bpython = "^0.22.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
